+++ tools/data/src/commands/mod.rs
@@ -1,3 +1,4 @@
+pub mod convert;
 pub mod count;
 pub mod head;
 pub mod interleave;