+++ CHANGELOG.md
@@ -19,6 +19,7 @@
 
 * We now always report a line immediately before printing the best move
 * The 'genfens' UCI command is now implemented to support OpenBench datagen
+* tools/data now has a command to convert from .pgn to viriformat
 
 ## [10.0]